modifications
deplacement des classes pour une meilleure structure de fichier

ajouts des commentaires de class et fonction
