+++ Source/CodeNameHippie/app/src/main/java/com/pam/codenamehippie/ui/MapsActivity.java
@@ -25,7 +25,6 @@ import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.pam.codenamehippie.R;
-import com.pam.codenamehippie.ui.adapter.CarteListeOrganismeAdapter;
 import com.pam.codenamehippie.ui.adapter.CarteOrganismeAdapter;
 import com.pam.codenamehippie.ui.view.trianglemenu.TestDonneeCentre;
 import com.sothree.slidinguppanel.SlidingUpPanelLayout;
@@ -36,7 +35,7 @@ import java.util.List;
 public class MapsActivity extends HippieActivity implements OnMapReadyCallback, ExpandableListView.OnGroupClickListener {
     private SlidingUpPanelLayout slidingLayout;
     private ExpandableListView expandableListView;
-    private ArrayList<Organisme> listOrganisme=new ArrayList<>();
+    private ArrayList<TestDonneeCentre.Organisme> listOrganisme=new ArrayList<>();
     private int ordre;
     private Intent intent;
     GoogleMap mMap;
@@ -147,7 +146,7 @@ public class MapsActivity extends HippieActivity implements OnMapReadyCallback, 
      *
      * @param listOrganisme
      */
-    private void prepareMarkers(final ArrayList<Organisme> listOrganisme, final int viewID) {
+    private void prepareMarkers(final ArrayList<TestDonneeCentre.Organisme> listOrganisme, final int viewID) {
         if(mMap!=null){
             mMap.clear();
         }
@@ -177,7 +176,7 @@ public class MapsActivity extends HippieActivity implements OnMapReadyCallback, 
                                                       break;
                                                   }
                                               }
-                                              final Organisme mOrganisme = listOrganisme.get(ordre);
+                                              final TestDonneeCentre.Organisme mOrganisme = listOrganisme.get(ordre);
                                               if(viewID!=R.id.main_organisme_image){
                                               expandableListView.setAdapter(new CarteOrganismeAdapter(MapsActivity.this, mOrganisme, viewID));}
                                               else{
